+++ app/tests/Invoice/InvoiceServiceVatTest.php
@@ -32,6 +32,7 @@ use App\Vat\CnbExchangeRateClient;
 use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Framework\Attributes\AllowMockObjectsWithoutExpectations;
 use PHPUnit\Framework\TestCase;
+use Psr\EventDispatcher\EventDispatcherInterface;
 
 #[AllowMockObjectsWithoutExpectations]
 final class InvoiceServiceVatTest extends TestCase
@@ -91,6 +92,7 @@ final class InvoiceServiceVatTest extends TestCase
             new IssuerProfileProvider($settings, new TaxProfileConfig($settings)),
             $this->createMock(IncomeUpserter::class),
             new DepositConfig($settings),
+            $this->createMock(EventDispatcherInterface::class),
         );
     }